#133 - reordenacao da estrutura do projeto para uma melhor escalabilidade
